requirements.txt
weasyprint
jinja2
python-dotenv